processor/src/main/java/dz/jtsgen/processor/mapper/name/UpperCamelCaseNameMapper.java
/*
 * Copyright (c) 2018 Dragan Zuvic
 *
 * This file is part of jtsgen.
 *
 * jtsgen is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * jtsgen is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with jtsgen.  If not, see http://www.gnu.org/licenses/
 *
 */

package dz.jtsgen.processor.mapper.name;

class UpperCamelCaseNameMapper implements NameMapper {

    @Override
    public String mapMemberName(String rawName) {
        if (rawName == null || rawName.length() == 0) return rawName;
        return rawName.substring(0,1).toUpperCase()
                + ( rawName.length() > 1 ? rawName.substring(1): "" );
    }
}
